Await ConnectionInfoReq before initializing project.
All other messages are buffered until ConnectionInfoReq has been
handled. The object is to alleviate a race condition in clients where
the protocol version is required to send an rpc, but the connection
info hasn't been received yet. NOTE: This PR does not eliminate the race,
but should make it much lower probability. The next step is to make
the connection info request blocking on the client.